backend/requirements.txt
flask>=3.1.3
flask-cors
python-dotenv
langchain
langchain-community
langchain-ollama
langchain-google-genai
faiss-cpu
pypdf>=6.7.2
tiktoken
pytest
